backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-redis==6.0.0
django-phonenumber-field[phonenumberslite]==8.4.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg==3.3.2
psycopg-binary==3.3.2
redis==7.2.0
sqlparse==0.5.5
tzdata==2025.3
waitress==3.0.2
whitenoise==6.11.0
